% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiexplorer-package.R
\docType{package}
\name{haiexplorer-package}
\alias{haiexplorer}
\alias{haiexplorer-package}
\title{haiexplorer: Interactive HAI Burden Explorer}
\description{
Data and tools to explore healthcare-associated infections (HAIs) using cases, deaths and DALYs. Includes a Shiny app to interactively explore the dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ETC5523-2025/assignment-4-packages-and-shiny-apps-serl0001/}
  \item \url{https://github.com/ETC5523-2025/assignment-4-packages-and-shiny-apps-serl0001/}
  \item Report bugs at \url{https://github.com/ETC5523-2025/assignment-4-packages-and-shiny-apps-serl0001/issues}
}

}
\author{
\strong{Maintainer}: Sandhya Erland Chandrasekar \email{serl0001@student.monash.edu}

}
\keyword{internal}
